requirements.txt
Flask
Flask-RESTful
Flask-JWT
FLASK-SQLAlchemy
uwsgi
psycopg2